+++ src/Plotter/Cellmap.h
@@ -9,17 +9,17 @@ namespace Plotter
 class Cellmap : public Heatmap
 {
   protected:
-	double process_cell(Range<DataNorm> x, Range<DataNorm> y, const DataSets::DataSet &data) const override
+	float process_cell(Range<DataNorm> x, Range<DataNorm> y, const DataSets::DataSet &data) const override
 	{
 		/* search for the two closest data points */
-		double cell_value = 0.0;
-		double minDistance[2]{ 2.0, 0.0 };
+		float cell_value = 0.0;
+		float minDistance[2]{ 2.0, 0.0 };
 		for (auto it = data.begin(); it != data.end(); it++)
 		{
 			auto dataPoint = *it;
-			double dx = pow(dataPoint.x() - x.mid(), 2);
-			double dy = pow(dataPoint.y() - y.mid(), 2);
-			double distance = sqrt(dx + dy);
+			float dx = pow(dataPoint.x() - x.mid(), 2);
+			float dy = pow(dataPoint.y() - y.mid(), 2);
+			float distance = sqrt(dx + dy);
 			if (distance < minDistance[0])
 			{
 				minDistance[1] = minDistance[0];
@@ -30,7 +30,7 @@ class Cellmap : public Heatmap
 		return cell_value;
 	}
 
-	void merge_distance(double &cell_value, double distance) const override
+	void merge_distance(float &cell_value, float distance) const override
 	{
 		/* Clip distance using the treshold */
 		if (distance > CELLMAP_TRESHOLD)
